track the config file line number of actions.
include the line number in the log message for any action.

for accepted and quarantined messages, insert a
X-Milter-Regex-Decision-Trace header in unauthenticated non-local
connections, giving the config file mtime epoch and line number of the
deciding action.  this, and the X-GeoIP-Summary, should probably be
togglable on the command line and/or with config file extensions.